Removes unneeded gems
