+++ Gemfile
@@ -1,6 +1,4 @@
 source 'https://rubygems.org'
 
 # Specify your gem's dependencies in rushb.gemspec
-gem 'rocket_pants', git: 'https://github.com/corneldm/rocket_pants'
-
 gemspec
