+++ lib/rushb.rb
@@ -1,20 +1,7 @@
-require "rocket_pants"
-require "api_smith"
+require 'httparty'
 
 require "rushb/version"
-
-require "rushb/smash/game"
-require "rushb/smash/match"
-require "rushb/smash/ranking"
-require "rushb/smash/player"
-
-require "rushb/resources/game"
-require "rushb/resources/player"
-require "rushb/resources/match"
-
 require "rushb/client"
-
-require "rushb/v2/client"
 
 module Rushb
 end
