+++ lib/rushb/version.rb
@@ -1,3 +1,3 @@
 module Rushb
-  VERSION = "0.0.1"
+  VERSION = "0.1.0"
 end
