+++ rushb.gemspec
@@ -20,8 +20,9 @@ Gem::Specification.new do |spec|
 
   spec.add_development_dependency "bundler", "~> 2.0.2"
   spec.add_development_dependency "rake", "~> 10.0"
-  spec.add_development_dependency "rspec"
-  spec.add_development_dependency "pry"
-  spec.add_development_dependency "vcr"
-  spec.add_development_dependency "webmock"
+  spec.add_development_dependency "rspec", "~> 3.8.0"
+  spec.add_development_dependency "vcr", "~> 5.0.0"
+  spec.add_development_dependency "webmock", "~> 3.7.6"
+
+  spec.add_dependency "httparty", "~> 0.17.1"
 end
