+++ spec/rushb/client_spec.rb
@@ -2,13 +2,13 @@ require 'rspec'
 require 'rushb'
 
 RSpec.describe Rushb::Client do
-  let!(:client) { Rushb::V2::Client.new("xxx") }
+  let!(:client) { Rushb::Client.new("xxx") }
 
   describe "#sending requests" do
     it "tries to send the request" do
       expect {
         client.games
-      }.to raise_error(Rushb::V2::ClientError)
+      }.to raise_error(Rushb::ClientError)
     end
   end
 end
