+++ spec/rushb/resources/game_spec.rb
@@ -3,7 +3,7 @@ require 'rushb'
 require_relative '../../support/vcr.rb'
 
 RSpec.describe Rushb::Resources::Game, type: :client, vcr: true do
-  let!(:client) { Rushb::V2::Client.new("4cdb7f00862e0133005b033abce37ff2") }
+  let!(:client) { Rushb::Client.new("4cdb7f00862e0133005b033abce37ff2") }
 
   describe '#games' do
     it 'returns with the games?' do
